relocate app into ui folder
